services/scoring-python/scoring_server.py
"""
IRONCLAD Risk Scoring and Failure Grammar Learner
Computes multi-axis risk scores and learns deployment failure patterns.
"""

import os
from flask import Flask, jsonify, request
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

@app.route('/health', methods=['GET'])
def health():
    return jsonify({"status": "healthy", "service": "scoring-python"}), 200

@app.route('/api/v1/score', methods=['POST'])
def score_deployment():
    """
    Score a deployment across blast radius, reversibility, and timing risk.
    Stub implementation - full scoring engine coming.
    """
    data = request.get_json()
    return jsonify({
        "blast_radius_score": 0.0,
        "reversibility_score": 0.0,
        "timing_risk_score": 0.0,
        "decision": "PENDING",
        "message": "Risk scorer not yet implemented"
    }), 200

@app.route('/api/v1/failure-grammar', methods=['GET'])
def get_failure_grammar():
    """Retrieve learned failure patterns for dashboard visualization."""
    return jsonify({
        "patterns": [],
        "message": "Failure grammar learner not yet implemented"
    }), 200

if __name__ == '__main__':
    port = os.getenv('SCORING_PORT', '8083')
    app.run(host='0.0.0.0', port=int(port), debug=True)
