services/scoring-python/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.24.3
scikit-learn==1.3.0
pydantic==2.4.2
pytest==7.4.3
